Calext rfc

+++ src/test/java/it/cnr/iit/jscontact/tools/test/converters/vcard2jscontact/VCardTest.java
@@ -418,9 +418,9 @@ public class VCardTest extends VCard2JSContactTest {
         assertEquals("testCompleteVCard6 - 81", "ez-vcard 0.9.14-fc", jsCard.getProdId());
         assertEquals("testCompleteVCard6 - 82", 1, jsCard.getAnniversaries().size());
         assertNotNull("testCompleteVCard6 - 83.a", jsCard.getAnniversaries().get("ANNIVERSARY-1").getDate().getPartialDate());
-        assertEquals("testCompleteVCard6 - 83.b", (long) 2016, (long) jsCard.getAnniversaries().get("ANNIVERSARY-1").getDate().getPartialDate().getYear());
-        assertEquals("testCompleteVCard6 - 83.c", (long) 8, (long) jsCard.getAnniversaries().get("ANNIVERSARY-1").getDate().getPartialDate().getMonth());
-        assertEquals("testCompleteVCard6 - 83.d", (long) 1, (long) jsCard.getAnniversaries().get("ANNIVERSARY-1").getDate().getPartialDate().getDay());
+        assertEquals("testCompleteVCard6 - 83.b", 2016, (long) jsCard.getAnniversaries().get("ANNIVERSARY-1").getDate().getPartialDate().getYear());
+        assertEquals("testCompleteVCard6 - 83.c", 8, (long) jsCard.getAnniversaries().get("ANNIVERSARY-1").getDate().getPartialDate().getMonth());
+        assertEquals("testCompleteVCard6 - 83.d", 1, (long) jsCard.getAnniversaries().get("ANNIVERSARY-1").getDate().getPartialDate().getDay());
         assertTrue("testCompleteVCard6 - 84", jsCard.getAnniversaries().get("ANNIVERSARY-1").isBirth());
         assertTrue("testCompleteVCard6 - 85", jsCard.getSpeakToAs().isMasculine());
